Support blocking user/guilds

+++ PyDiscordBot/misc/Events.py
@@ -5,10 +5,24 @@ from PyDiscordBot.utils import DataUtils
 
 async def after_invoke(ctx):
     if DataUtils.guilddata(ctx.guild.id).get('deleteCommand') is True:
-        try:
-            await ctx.message.delete()
-        except:
-            pass
+        if ctx.command.module != 'PyDiscordBot.commands.Owner':
+            try:
+                await ctx.message.delete()
+            except:
+                pass
+
+
+async def before_invoke(message):
+    try:
+        if DataUtils.blockeddata(message.author.id).get('state'):
+            return False
+        if message.guild is not None:
+            if DataUtils.blockeddata(message.guild.id).get('state'):
+                return False
+        else:
+            return True
+    except AttributeError:
+        return True
 
 
 class Events(commands.Cog):
@@ -20,11 +34,13 @@ class Events(commands.Cog):
         toinsert = [
             {"_id": guild.id, "guild_id": guild.id, "modlog_status": "NONE", "deleteCommand": True}
         ]
-        DataUtils.database().insert_many(toinsert)
+        DataUtils.guild_database().insert_many(toinsert)
 
     @commands.Cog.listener()
-    async def on_command(self, ctx):
-        await after_invoke(ctx)
+    async def on_message(self, message):
+        if await before_invoke(message):
+            await self.bot.process_commands(message)
+            await after_invoke(await self.bot.get_context(message))
 
 
 def setup(bot):
